requirements.txt
asttokens==2.0.8
backcall==0.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
decorator==5.1.1
executing==1.1.1
idna==3.4
ipython==8.5.0
jedi==0.18.1
matplotlib-inline==0.1.6
numpy==1.23.4
opencv-python==4.6.0.66
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
prompt-toolkit==3.0.31
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.13.0
PyGObject==3.26.1
python-apt==1.6.5+ubuntu0.7
requests==2.28.1
setuptools-git==1.2
six==1.16.0
stack-data==0.5.1
torch==1.12.1
torchvision==0.13.1
tqdm==4.64.1
traitlets==5.5.0
typing_extensions==4.4.0
unattended-upgrades==0.1
urllib3==1.26.12
wcwidth==0.2.5
